Changes made -> - Rebuilt the logic of areamanager page - changed names of some files
