+++ src/main/resources/org/lenden/area_manager.fxml
@@ -1,76 +1,79 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.control.ScrollPane?>
-<?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.BorderPane?>
 <?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.TilePane?>
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" minHeight="-Infinity" minWidth="-Infinity" prefHeight="699.0" prefWidth="1222.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.lenden.TableSettingsController">
-   <children>
-      <BorderPane maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" minHeight="500.0" minWidth="500.0" prefHeight="699.0" prefWidth="1222.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
-         <bottom>
-            <ScrollPane fitToHeight="true" fitToWidth="true" prefHeight="594.0" prefWidth="1202.0" BorderPane.alignment="CENTER">
-               <BorderPane.margin>
-                  <Insets bottom="10.0" left="10.0" right="10.0" />
-               </BorderPane.margin>
-               <content>
-                  <AnchorPane>
-                     <children>
-                        <TilePane fx:id="tableAndAreaTilePane" alignment="CENTER" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" style="-fx-background-color: white; -fx-border-color: grey; -fx-border-width: 0.5; -fx-border-radius: 5;" vgap="10.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
-                           <padding>
-                              <Insets top="15.0" />
-                           </padding>
-                        </TilePane>
-                     </children>
-                  </AnchorPane>
-               </content>
-            </ScrollPane>
-         </bottom>
-         <top>
-            <HBox alignment="CENTER" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" minHeight="150.0" minWidth="800.0" prefWidth="1201.0" spacing="150.0" style="-fx-border-color: grey; -fx-border-radius: 5px;" BorderPane.alignment="CENTER">
+<BorderPane xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.lenden.AreaManagerController">
+   <top>
+      <HBox alignment="CENTER" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" minHeight="150.0" minWidth="800.0" prefWidth="1201.0" spacing="150.0" style="-fx-border-color: grey; -fx-border-radius: 5px; -fx-background-color: white;" BorderPane.alignment="CENTER">
+         <children>
+            <VBox fillWidth="false">
                <children>
-                  <VBox fillWidth="false">
-                     <children>
-                        <Label fx:id="totalAreasLabel" alignment="CENTER" prefHeight="91.0" prefWidth="163.0" text="6">
-                           <font>
-                              <Font size="58.0" />
-                           </font>
-                        </Label>
-                        <Label text="Total Areas">
-                           <font>
-                              <Font name="Segoe UI Semibold" size="32.0" />
-                           </font>
-                        </Label>
-                     </children>
-                  </VBox>
-                  <VBox fillWidth="false">
-                     <children>
-                        <Label fx:id="totalTablesLabel" alignment="CENTER" prefHeight="91.0" prefWidth="175.0" text="35">
-                           <font>
-                              <Font size="58.0" />
-                           </font>
-                        </Label>
-                        <Label text="Total Tables">
-                           <font>
-                              <Font name="Segoe UI Semibold" size="32.0" />
-                           </font>
-                        </Label>
-                     </children>
-                  </VBox>
+                  <Label fx:id="totalAreasLabel" alignment="CENTER" prefHeight="91.0" prefWidth="163.0" text="6">
+                     <font>
+                        <Font size="58.0" />
+                     </font>
+                  </Label>
+                  <Label text="Total Areas">
+                     <font>
+                        <Font name="Segoe UI Semibold" size="32.0" />
+                     </font>
+                  </Label>
                </children>
+            </VBox>
+            <VBox fillWidth="false">
+               <children>
+                  <Label fx:id="totalTablesLabel" alignment="CENTER" prefHeight="91.0" prefWidth="175.0" text="35">
+                     <font>
+                        <Font size="58.0" />
+                     </font>
+                  </Label>
+                  <Label text="Total Tables">
+                     <font>
+                        <Font name="Segoe UI Semibold" size="32.0" />
+                     </font>
+                  </Label>
+               </children>
+            </VBox>
+         </children>
+         <padding>
+            <Insets bottom="40.0" top="40.0" />
+         </padding>
+         <BorderPane.margin>
+            <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+         </BorderPane.margin>
+      </HBox>
+   </top>
+   <bottom>
+      <ScrollPane fitToHeight="true" fitToWidth="true" prefHeight="594.0" prefWidth="1202.0" style="-fx-background-color: white;" BorderPane.alignment="CENTER">
+         <BorderPane.margin>
+            <Insets bottom="10.0" left="10.0" right="10.0" />
+         </BorderPane.margin>
+         <content>
+            <VBox fx:id="areasVbox" alignment="TOP_CENTER" prefHeight="200.0" prefWidth="100.0" spacing="15.0" style="-fx-background-color: white; -fx-border-color: grey; -fx-border-radius: 5px;">
                <padding>
-                  <Insets bottom="40.0" top="40.0" />
-               </padding>
-               <BorderPane.margin>
                   <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
-               </BorderPane.margin>
-            </HBox>
-         </top>
-      </BorderPane>
-   </children>
-</AnchorPane>
+               </padding>
+               <children>
+                  <Label fx:id="unsavedChangesLabel" text="You have un-saved Changes *" textFill="#a80101" visible="false" />
+                  <HBox alignment="CENTER" prefHeight="46.0" prefWidth="1178.0" spacing="20.0">
+                     <children>
+                        <Button minHeight="38.0" mnemonicParsing="false" onMouseClicked="#saveButtonListener" prefHeight="38.0" prefWidth="150.0" styleClass="saveEdit-button" stylesheets="@../../settingsStyleSheet.css" text="Save " />
+                        <Button minHeight="38.0" mnemonicParsing="false" onMouseClicked="#addAreaButtonListener" prefHeight="38.0" prefWidth="150.0" styleClass="saveEdit-button" stylesheets="@../../settingsStyleSheet.css" text="Add Area" />
+                     </children>
+                     <VBox.margin>
+                        <Insets bottom="20.0" />
+                     </VBox.margin>
+                  </HBox>
+               </children>
+            </VBox>
+         </content>
+      </ScrollPane>
+   </bottom>
+</BorderPane>
